Move original regex parser/evaluator into a nested package.
