+++ scala-exercises-server/app/controllers/ExercisesController.scala
@@ -3,7 +3,7 @@ package controllers
 import models.ExerciseEvaluation
 import play.api.libs.json.{ JsError, JsSuccess, Json }
 import play.api.mvc.{ Action, BodyParsers, Controller }
-import services.parser.ExercisesService
+import services.ExercisesService
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
