+++ scala-exercises-server/app/services/exerciseV0/ExercisecodeExtractor.scala
@@ -1,4 +1,4 @@
-package services.parser
+package services.exerciseV0
 
 import models.{ Section, Category, Exercise }
