+++ scala-exercises-server/app/services/exerciseV0/ExercisesService.scala
@@ -1,4 +1,4 @@
-package services.parser
+package services.exerciseV0
 
 import java.io.File
 import java.net.URLClassLoader
